requirements.txt
# Trading & Market Data
alpaca-py>=0.21.0
ccxt>=4.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# Technical Indicators
ta>=0.11.0

# Scheduling
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0
aiohttp>=3.9.0

# Deep Learning
torch>=2.0.0
scikit-learn>=1.3.0
yfinance>=0.2.0

# Logging
rich>=13.0.0
joblib>=1.3.0
